Refactor some modules for backwards compatibility
