+++ labrea/application.py
@@ -13,7 +13,7 @@ from typing import (
 )
 
 from .arguments import Arguments, arguments
-from .evaluatable import Evaluatable, MaybeEvaluatable, Options
+from .types import Evaluatable, MaybeEvaluatable, Options
 
 P = ParamSpec("P")
 A = TypeVar("A", covariant=True)
