+++ tests/test_arguments.py
@@ -1,7 +1,7 @@
 import pytest
 
 from labrea.arguments import Arguments, EvaluatableArgs, EvaluatableKwargs, arguments
-from labrea.evaluatable import KeyNotFoundError
+from labrea.exceptions import KeyNotFoundError
 from labrea.option import Option
